API

API

Details about Psycho's functions.

Core

Psycho.standardize — Function.

standardize(X; robust::Bool=false)

Standardize (scale and reduce, Z-score) X so that the variables are expressed in terms of standard deviation (i.e., mean = 0, SD = 1).

Arguments

X: Array or DataFrame.

robust::Bool: If true, the standardization will be based on median and mad instead of mean and sd (default).

Note

Ideas / help required:

Deal with missing values (See #4)

Examples

julia> standardize([1, 2, 3])
3-element Array{Float64,1}:
 -1.0
  0.0
  1.0

source

Simulate

Data

Psycho.simulate_data_correlation — Method.

simulate_data_correlation(coefs::Vector{<:Number}; n::Int=100, noise::Number=0.0)

Generate a DataFrame of correlated variables.

Arguments

coefs::Vector{<:Number}: Correlation coefficients.

n::Int: Number of observations.

noise::Number: The SD of the random gaussian noise.

Note

Ideas / help required:

Different group sizes (See #9)

Examples

julia> simulate_data_correlation([0.2])
100×2 DataFrames.DataFrame
[...]

source

Misc

Psycho.simulate_coefs_correlation — Function.

simulate_coefs_correlation(coefs_mean::Number=0.1; coefs_sd::Number=0.1, n::Int=10)

Generate a vector of random correlation coefficients from a normal distribution.

Arguments

coefs_mean::Number: Mean of the normal distribution from which to get the coefs.

coefs_sd::Number: SD of the normal distribution.

n::Int: Number of coefficients.

Examples

julia> simulate_coefs_correlation(0.5)
10-element Array{Array{Float64,1},1}:
[...]

source

Psycho.simulate_groupnames — Method.

simulate_groupnames(n::Int; nchar::Int=2)

Create vector of random group names of length n containing nchar characters.

Arguments

n::Int: Number of group names.

nchar::Int: Number of random characters in the name.

Note

Ideas / help required:

Can be enhanced to make it more procedural and less random (See #8)

implement different types simulations (e.g., "A, B... AA, AB...") (See #8)

Examples

julia> simulate_groupnames(10)
10-element Array{String,1}:
[...]

source