API

API

Details about Psycho's functions.

Core

Psycho.standardize — Function.

standardize(X; robust::Bool=false)

Standardize (scale and reduce, Z-score) X so that the variables are expressed in terms of standard deviation (i.e., mean = 0, SD = 1).

Arguments

X: Array or DataFrame.

robust::Bool: If true, the standardization will be based on median and mad instead of mean and sd (default).

Note

Ideas / help required:

Deal with missing values

Examples

julia> standardize([1, 2, 3])
3-element Array{Float64,1}:
 -1.0
  0.0
  1.0

source

Simulate

simulate_groupnames(n::Int; nchar::Int=2)